Fix broken redirect on GraphQL 403
Update `useAuthenticatedFetch` to match the updated template[1]
to include the fix from Shopify/shopify-frontend-template-react#66.

See linked PR for more details.

Note that the experience is still a little broken: on a 403, we
redirect to the auth flow, lose the user input and their current page,
we end up landing on the app's home page.

But at least that's better than the current state of things.

[1]: https://github.com/Shopify/shopify-frontend-template-react/blob/a49d4aa7d1f425e32b8440efd63e00519b321af8/hooks/useAuthenticatedFetch.js